Eng 13597 (#5182)
ENG-13597: Fixing 2 bugs:
1. On recover, push from EE could trigger pollImpl before truncation is called on the export source PBD.
Truncation after those polls could delete the segment that was polled. To fix this,
made changes to ExportDataSource so that it keeps track of whether it is ready to start polling or not.
2. If readyForData is called as a result of catalogupdate from command log replay, it could trigger poll.
This was because readyForData was called with false, which would directly call buildListener without checking
the barrier. To fix this, we now simply use a flag that keeps track of whether startPolling was called or not.
This can be used in both initialize and updateCatalog path.