ENG-13701: (#5199)
There are two maps which track the set of blocks that are non-empty. These maps are used to determine if a block is in a bucket. However, an optimization to keep a cache block around when the table's tuple count goes to zero does not cleanly track this buffer either when the last tuple is deallocated nor subsequently when the first tuple is allocated. Therefore, this block may not participate in subsequent compactions.

This branch adds special checks to ensure that the last empty block is correctly tracked when the last tuple is deleted and when this block is used by the first tuple that is subsequently inserted.