Eng 13694 Shared Replicated Table View Crash (#5210)
* Force using the right top end.

Sometimes the lowest numbered thread does work on behalf of other
threads, and it uses the ExecutorContext of these other threads
to do that work.  This causes hilarity when the progress monitor
proxy reports progress on a thread operation on which the server
is not expecting.

* Move thread local data to the ExecutorContext class.

This is parallel to the pthread_specific key static_key.

* Change to make the lowest site's top end thread specific and static.

We keep a pointer to the lowest numbered site's top end thread
specific and static.  This allows us to communicate only with the
lowest numbered site's top end when reporting progress to java.

* Merge remote-tracking branch 'remotes/origin/master' into ENG-13677-C6MemcheckFails

* Merge remote-tracking branch 'remotes/origin/master' into ENG-13677-C6MemcheckFails

* Change some comments.