ENG_13769, fix a rare window that in node rejoin we may forward messages included in stream snapshot to rejoin site.
In duplicate counter map, the assumption that every message is ordered by txnId is wrong, snapshot save fragment message
has MP txnId, in SP heavy workload the MP sequence number is usually smaller than SP sequence number, which ends up leaving
the fragemnt message in the first entry of duplicate counter map. In a word, the actual order may vary so we can't reply on
the relative order of snapshot fragment to know which messages to forward.

So instead of just looking at MP txnId, we first find the actual fragment in duplicate counter map, then use the spHandle to
filter out messages we want to forward, exclude the snapshot fragment message itself.

Change-Id: Idef33f1595a5286a06c829d5d88b572ecdbfff48