ENG-13769: (#5266)
The dump log generated after a MP deadlock logs the transactiontaskqueue twice. Also some of the dump messages are written to the tmlog while others are written to the host log. Finally, the timestamp added to restarted fragments and completions is encoded so we print out the details so it is easy to figure out which restart this is.