ENG-13769: The Repair/Restart flag in the seq number was incorrectly setting bit 20 instead of 22 but looking for bit 22 in the check.
The Scoreboard was using a hashmap on the taskqueue but is now using an arraylist indexed by siteId which makes the debug output of the scoreboard cleaner (at the cost of slightly more complicated initialization sequence).

There is a slight possibility for stale Completion from a dead MPI to arrive after a new MPI repair message so now we check for stale initial completions specifically.