ENG-13769: (#5269)
* ENG-13769:
Some of the unit tests were failing because the static in the scoreboard was being reused by local serverthread. Also the RestartSequence generator uses the zkzeeper node id of the MPI to create unique sequence numbers, but some unit tests use mock which does not expose the nodeId.

* ENG-13769:
Fixed size fields in FragementTaskMessage should always preceed variable sized fields because the size of an initiate task in the fragment task depends on it being last.

* ENG-13769:
Don't skip fragment check when a mismatched completion arrives. Also use all the bits in the timestamp, and convert the int to a long before the shift left.