Eng 13769+12628: ScoreBoard support for NP transaction and JoinNode SiteId fix (#5279)
* ENG-13769:
The scoreboard now coordinates all fragment tasks across all sites. However, NP transactions such as @BalancePartitions does not need to be coordinated because it does not modify replicated tables. This change will resolve join failures that currently get blocked by the scoreboard.

* ENG-13769:
Fix some unit tests and use an array instead of an arraylist for the scoreboard so that contiguous siteIds higher then 0 (happens when local serverthread is reused) are tolerated.

* ENG-12628: Pospone ElasticJoinProducer Mailbox initilzation till sp
initors creation

* ENG-12628:
MP Repair algo creates a new timestamp every time the repair is restarted so the same timestamp will be used repeatedly. We now create the timesamp once in the mpinitiator and pass it in every time we create a new MP Repair algo.

* remove duplicated kFactor method because of rebase

* address the review