Eng 13698 jet brains metadata (#5273)
    We want to support CLion and IntelliJ projects, for C++ and Java
development respectively. The two problems are that these projects
cannot both reside in the same directory, and the CLion project must
reside in the directory which contains the main CMakeLists.txt file. The
latter means that the CLion project has to reside in the top level
directory, and the IntelliJ project must reside some place else.

    We put the CLion files in the root of the voltdb community
hierarchy. We put the IntelliJ files in a directory named
"tools/IntelliJ". The metadata files are found in a directory called
.idea for both CLion and IntelliJ.

    So, to open an IntelliJ project one just needs to open
voltdb/tools/IntelliJ. The .idea file will be there, and if the pro
directory exists at voltdb/tools/IntelliJ/../../../pro, it will be
picked up. To open a CLion project one needs to open the project at
voltdb.

    Note that the pro directory needs to be a sibling of voltdb and
needs to be named "pro". If this is a problem we can try to fix this.