ENG-13894: (#5319)
Load table was not generating consistent results on all sites in the event of a failure. While this failure should not happen, it would generate a Hash Mismatch if it does, so this code fixes this path and addes unit tests for both this failure scenario and the exception failure scenario.