ENG-13854: (#5341)
* ENG-13854:
Rename restartable in the completion message to indicate that this is a repair completion with an abort.

Also use and advance the timestamp for this use of completions.

Also make sure that abort completions for txns where some sites have the first fragment and some don't, mark the txn state complete in those sites that receive the txn. Otherwise the backlog.flush won't advance the backlog.

Clean up unit tests that flush the backlog at the time the txn is added rather than waiting until the txn has been released by the scoreboard.

* ENG-13854:
Improve comments. Remove unnecessary imports in test. Remove Commnented out code.