ENG-13880: In order to migrate an SPI on a DR Consumer, the BufferReceivers associated with that Consumer must be migrated as well. The first step in the process is to shutdown the PartitionBufferReceivers pinned to that consumer partition.
While the shutdown process is straightforward, the problem is that the DRConsumerMpCoordinator may not have any transaction in process from this PartitionBufferReceiver. To enable this a new mechanism has been added to reneg (take back) a pending message from the DRConsumerMpCoordinator. A response will be provided by the DRConsumerMpCoordinator when either, the message (if any) was successfully reneged, or the transaction associated with the message (and is currently in progress), has been completed.