Eng 14165 sqlcov internal err (#5414)
ENG-14165: modified SqlCoverage to fail when "unexpected internal error occurred" (and certain other errors) is returned by VoltDB; the number of such errors is now included in the HTML table of results; and the types of ('Fatal' or 'Non-fatal') Exceptions/Errors found is printed on the console. Also, labeled old & deprecated config files as such, mostly by adding 'old_' prefix to file names or, in one case, via a comment.