ENG-14292: (#5468)
The previous boolean logic was correct but obscure because an arriving MP Read with the same TxnId as the MP Read TxnId would evaluate correctly only because the set of != txnIds is a superset of > txnIds:
          if ((txnState.isReadOnly()) && headTask.getTransactionState().isReadOnly() &&
                   TxnEgo.getSequence(task.getTxnId()) != TxnEgo.getSequence(headTask.getTxnId())
                      || (TxnEgo.getSequence(task.getTxnId()) > TxnEgo.getSequence(headTask.getTxnId()))) {
While the expression could semantically resolve (true && true && false || true) to true, in practice if the third term is false, the fourth term must be false. However, the intent of the code is more precise with this change.