ENG-14262: Add more clarity to the Dump messages by: 1. removing raw txnIds. 2. including pending MP txns in the SP Site output to identify MP deadlocks because of blocked MP Txns. 3. Including a list of MPProcedure names in the MP site backlog so that snapshots or UACs that look like they are blocked but are in fact just pending can be identified more easily.
There is also some message check reordering in the deliverInternal path of the InitiatorMailbox because the most common messages (InitiateTask and FragmentTask) should really be checked first rather than last. The only reason for the existing order would be if there were subclasses of these messages that require special handling, but that is not the case.