Eng 14251 Better Planner Testing Framework (#5469)
Improve the plan tester functions.  We introduce a function, validatePlan, and some
ancillary functions to describe an expected fragment or fragments. The function validatePlan
takes a SQL string and one or two fragment descriptions, compiles the SQL and compares
the resulting fragments with the expected ones.

These kinds of tests are much easier to understand than the existing plan dissection
tests.