Eng 14262 block dead connection before repair (#5511)
* ENG-14262:
There is a race between intra-cluster connections being closed and the beginning of a Repair. If messages arrive after the repair and this host is the only one to receive it, other hosts could be out of sync. This is particularly true when the MPI initiates a repair that cleans up the scoreboard. Inaccurate repairs can cause MP Deadlocks in hosts that receive fragments or completions that none of the other remaining hosts report in their repair logs.

The fix is to propagate the dead host in the repair message. If the dead host's connection is still present in hosts that receive the repair, the repair is deferred until the connection is closed.

* ENG-14262:
Track zookeeper and piconetwork host failures in seperate lists since they race to report.