ENG-14408: Use a thread to wait for timeout
The implementation of using a ScheduledThreadPoolExecutor with a core
pool size of 0 and keepAliveTime is 0 any thread in the pool will do
spin loop until the queue is empty. Instead of using a thread pool which
has no real defined start and end to its lifetime use a thread to track
handle tracking the timeout.

Create DelayedExecutionThread which wait for a delay to pass and then
invokes a runnable. The delayed thread can be canceled while waiting for
the delay to pass but not before.