ENG-14167 snapshot view hidden column (#5547)
In ENG-10892 - #5376, @MarcxJin removed the restriction of having a COUNT(*) column in single table materialized views. The way we did this is that we added a hidden column to the view tables to store the count. However, since we started to snapshot some materialized views in ENG-11578 - #5250, the lack of a way to save and restore those hidden columns to and from snapshots will always cause a failure. This pull request addresses this problem.