ENG-14447, fix rejoin timeout issue.
We used to shutdown SpTerm when the site is no longer leader, however during leader migration there is a short window between the new leader fails before leader migration comes to finish and the old leader accepts promotion. If rejoin happens in this window, SpTerm will fail to add the new site id into the replica list, it can cause rejoin node hangs on receiving stream snapshot. Because of this reason we keep the SpTerm even if it's no longer the leader.

Change-Id: Ia88327827fb1fcb4eb0084f75d48347f15005db3