Added a new enumerator to force large temp table queries. (#5537)
Added a new enumerator to force large temp table queries.

There is no test for this really.  The queries in TestOrderBySuite
don't actually work in large temp table mode.  But the test in
TestAdHocLarge runs and passes when we use the new enumerator,
and I can see that the plans are large temp table plans in the
TestGroupBySuite tests, even when we use "@AdHoc".  The tests
fail in the EE, though.