ENG-14444 - Process all logs in an MP transaction together
The binary logs of an MP transaction are split up into one log from each
producing partition and can usually be applied that way except when a
truncate is involved. With a truncate in each log the truncate will be
applied multiple times resulting in inconsistent tables.

Instead of applying each log separately in the ee send the ee all logs
and let it process the logs as it sees fit. This is done by prepending
the log count to the logs being sent down to the ee and then the ee will
process all logs at the same time.

One log will be applied until a truncate is reached when that occurs the
ee will switch to the next log read until the truncate. This will occur
until a truncate is read from each log. The truncate will then be
applied and log reading will pickup with the first log. This process
will keep occurring until the end of all logs are reached.