ENG-14335: tests of VMC's DR, Importer & Analysis tabs
Added, to NavigatePagesBasicTest.groovy, a checkVisibilityOfTabLinks
test that does just that, including the old tabs, the new-ish Analysis
tab (always visible), and the DR and Importer tabs, both of which may be
expected to be visible or not, depending on 2 System properties; also
new navigateDrPage, navigateImporterPage, and navigateDrAndImporterPage
tests, which run only if one or both of those System properties are set;
and added the Analysis tab to the navigateUsualPages test (formerly
navigatePages). In VoltDBManagementCenterPage.groovy, added new
isFOOLinkVisible methods, where FOO is: DbMonitor, Analysis, Admin,
Schema, SqlQuery, Dr, or Importer; to be used by the new
NavigatePagesBasicTest.checkVisibilityOfTabLinks test above; and tweaked
existing methods and links, mainly by ordering them in the same order as
they appear on the web page. Tweaked build.gradle and run.sh to include
the new 'dr' and 'importer' System properties. And added new <dr> and
<import> tags to deploy_pro.xml (the latter may need a bit of
adjustment).