First attempt at fixing the problem of matview fix and serial aggs for LTTs.
This does not work, but I think it's closer.  The problem now is that
there is a scan node of a view with columns A1, B1.  The scan node
projects onto just A1, but we later need to reaggregate onto A1 and
B1 both.  So we get a failure, which is quite correct.  This used to
work, so somehow the fix in this commit is not doing what the old code
did.  But I can't see it, and I've run out of time.  Sorry.