Eng 14335 vmc new tabs (#5577)
ENG-14335: tests of VMC's DR, Importer & Analysis tabs:
Added, to NavigatePagesBasicTest.groovy, a checkVisibilityOfTabLinks
test that does just that, including the old tabs, the new-ish Analysis
tab (always visible), and the DR and Importer tabs, both of which may be
expected to be visible or not, depending on 2 System properties; also
new navigateDrPage, navigateImporterPage, and navigateDrAndImporterPage
tests, which run only if one or both of those System properties are set;
and added the Analysis tab to the navigateUsualPages test (formerly just
navigatePages). In VoltDBManagementCenterPage.groovy, added new
isFOOLinkVisible methods, where FOO is: DbMonitor, Analysis, Admin,
Schema, SqlQuery, Dr, or Importer; to be used by the new
NavigatePagesBasicTest.checkVisibilityOfTabLinks test above; and tweaked
existing methods and links, mainly by ordering them in the same order as
they appear on the web page. Tweaked build.gradle and run.sh to include
the new 'dr' and 'importer' System properties and updated the latter to
include a -Pdr=true arg whenever server-pro is run, and added a comment
suggesting that -Pimporter=true should also be included, once ENG-14518
is fixed.
Added a new kafka_import_table to ddl.sql; and updated deploy_pro.xml to
refer to it in a new <import> tag, and added a new <dr> tag; also added
kafka_import_table to expectedTables.txt and expectedDefaultStoredProcs.txt;
and updated comments in those and related files.