ENG-14543: reduce space taken up by sqlcov *.html files (on Jenkins)
For SqlCoverage test suites with expected mismatches, do not produce any
'reproducer' files; and delete any specific query result files (assuming
the expected number of mismatches is not exceeded), to save space in
Jenkins.
Also slightly increased the number of expected mismatches for the
'joined-matview-default-full' test suite, based on a recent run; and
fixed a small issue where the number of columns printed was one too
small (in sql_coverage.test.py); and improved 'Key' text (in
SQLCoverageReport.py).