ENG-14622: Add setupProcedure method to the EE
Currently startTime is only cleared when executePlanFragment,
loadCatalog or updateCatalog is called. That means that if a procedure
runs in the EE and does not go through one of those paths it will have a
bogus start time. Rename setProcedureName to setupProcedure and update
it so that it reset startTime and logDuration for the procedure to be
executed. Also, add a completeProcedure to clear the procedure name.