debuglog: Perform logging in single macro
The debug logging was being split into three separate calls to printf.
This meant that if multiple logs were being done at the same time the
output could be all jumbled together. By moving everything into a single
macro there is only one call to printf which means that it will all be
printed together and you can not have one log being mixed with another.

Also, add milliseconds to the logging so that the logging timestamp
matches the java logging timestamp.