Eng 14614 exclude invocation from debug output on non site thread m (#5626)
* ENG-14614:
It is possible for the MP Deliver and the MP Site threads to debug print the Invocation for the same transaction at the same time. When this happens the parameters can be printed as well. It appears that some parameters (specifically VoltTable) are not thread safe at least for ToString(). To avoid this, only the TxnSummary will be printed out on non site threads and the invocation will only be printed out in detail on the site thread.

* Minor Cleanup

* ENG-14614:

Also protect Debug TM log output in MP Scheduler Deliver (which executes on the MPInitiator Deliver thread.