ENG-14602-14606: Remove table as interst during truncate
When a table is truncated it can register an undo action which can
result in calling the destructor on the current table. When a row is
deleted from a table the table from which the row was deleted is
registered as an UndoQuantumReleaseInterest. If a delete from a table
occurs before a truncate of that table that will cause the
notifyQuantumRelease() method to be invoked on a table after it had been
destroyed.