ENG-14637: (#5640)
It is possible to release or undo the same transaction twice if the repair completions are received on all sites before the CompletionReponse messages are processed on any of the sites. To ensure that we don't execute the completion twice we should check the transaction state of the transaction that releases the completion to make sure that it is not marked as DONE. If it is marked as DONE, the completion has already been processed and the execution of the transaction should be skipped.