Eng 12931: 6.5.x (#5670)
Eng 12931: poison pill in ee (#4844)

Generate a poison pill from the EE when the EE detects that there are 2 begin transactions without an end transaction. The poison pill will also print out the partial buffer that was generated so that the prior TXN that did not commit or rollback can be identified.

DR will be disabled on the producer when the poison pill is generated. However, if the producer cluster is reset, new consumers can reconnect and get a fresh snapshot.

- Fix up IPC path to include topend callback for pushPoisonPill

- Modified all other checks on DRTupleStream that throw execeptions except extendBuffer to generate a poison pill instead of throwing a fatal crash exception.

- Add Unit Test that verifies poison pill functionality.

- Improve code readability slightly. Also missed an early return in endTransaction() so that we don't start a new buffer after a failure with an endTransaction().