ENG-14347: As export streams build buffers, the number of records in buffers can diverge on different hosts due to the timing of the Tick() calls on the individual sites of each host. This in turn can result in different buffer counts for the same stream and partition on different hosts although the total number of records (sequence number) are the same. The first row of each buffer contains the streams schema, but this schema data is included in the USO calculation. Because we use stream offset (USO) to coordinate acknowledgements between hosts, the different number of buffers can result in the same record on different hosts starting at different USOs resulting in acking the wrong record. At promotion or SPI migration time, we can either skip past a record or send the same record twice with different USOs. Further, the different USO values can cause us to report that there are non-empty Export Streams when in fact all the Export data has been processed.
This fix removes the Schema out of the first record of the buffer allowing for consistent USO assignments for all records generated for the same stream on different hosts. The movement of the schema out of the record and into the buffer header is transparent to the connector interface.

A version field has also been added to the buffer header to disambiguate the format of the buffer for offline processing. The generation field has also been moved to the header so that buffers from identical generations can skip past the schema parsing step when decoding the buffer.