ENG-14162: (#5695)
A customer has a catalog that exceeds the 50MB message size. In most of the paths, catalogs and catalog diffs are passed around compressed but UAC passes the catalog diffs uncompressed to each host. This branch passes the catalog diffs compressed and asks each host to decompress locally before building the local jar file. This reduces the size of the message by half.