ENG-14774: Provide a consistent way to get list of new parttiions
ElasticJoinCoordinator was receiving the list of new parttions that were
to be added from Cartogropher and then expecting AbstractTopolgy to
generate a new topology with those partitions but nothing really
guaranteed that they would do the something.

Update AbstractTopology.mutateAddNewHosts to also return a list of the
new partition IDs. This allows ElasticJoinCoordinator to be consistnet
with regards to which partitions are being added and also where they
were added.