Temporarily removed exception (#5731)
in case of unregistered thread-local memory location.

Will systematically come back to thread-local memory
management in ENG-14906.

With release build, assert checking is voided, the
unchecked memory will be leaked. With debug build
(for system test and other Jenkins tests), the assertion
is in effect, and it will abort the database when this occurs;
When logging level is anything higher than logging disabled,
it will print stacktrace and die with assertion error.