Log a stacktrace as a single printf
Since stack traces are written as multiple printf statements their
printing could be interrupted by other log statements. To prevent this
from happening build a single string stack trace and log that in one
printf.

Reduce the implementations of rendering a stack trace down to one method
which operates on an ostream.