ENG-15083:
There is a race between UAC notifcations to export and EOS notifications from Sites. This means that if either is used to trigger the removal of the ExportDataSource, the other fails an assertion. This is exposed with the TestExportStatsSuite which is currently failing.

However, the m_eos flag is no longer used so the notification of the EndOfStream is no longer necessary. There is a small chance that the Export streams may start injecting events into the PBD. If this ends up being the case, we will still need the CPP api so the Topend callback will be kept until the need for this is decided.