TestCollector: Make all test logs relative to testing date
The test was always creating a log file for January 2 of this year and
December 31 and 27 of last year and then would collect the last 3 days of logs.
This would work most days except the first few days of the new year when those
hard coded dates would be in the three day range. The test for specifically
collecting over the year boundry had a similar issue. Instead of using any hard
coded dates pass in a date for now and make all log files relative to that date
so that the number of files being collected are consistent regardles of what
the current date is.