ENG-15130: If an MP Stored Procedure Txn contains a batch (voltExecuteSQL) that modifies a partitioned table before modifying a replicated table in the same batch and the partitioned table modification fails for any reason (usually a constraint violation but could also be a missing row), the result will be a Multi-Partition deadlock.
This is due to the coordination of the replicated table that requires all sites to be ready to work on the same statement at the same time (to ensure correct ordering and consistency). Since the failure on one site will break out the the execution, the other sites wait forever for the one site that gave up on the transaction.

The fix is to not break out of the transaction in the event of an exception but remember the exception and break out after all fragements have executed.