ENG-15166: Always consume partition data and replicas
The root cause of the issue was that there were two queues one for the partition
data and one for the children data. If an ignored exception was thrown
retrieving the partition data then the item in the children queue was not
removed which caused the two queues to become out of sync.

This change simplifies the async retrieval of partition information by making a
single utility method which returns a list AsynPartitions. This ties the async
request for the partition data to the child request so that the processing of
both will not get out of sync. Also, this condenses two different implementation
of this algorithm down to one common implementation.