LeaderElector: Simplify parts of the elector
m_done does not need to be an AtomicBoolean, since no compare and swap
operations are being performed it works just as well as a volatile boolean.
watchNextLowerNode() never returned the current leader it just returned this
node or the node which is ahead of it in the succession order. It is simpler to
have this method return a boolean indicating whether or not this node is the
leader.