OP-362: (#5867)
Sometimes PBD files that would normally be deleted need to truncated. However, the truncation does not occur if the file is already closed so this branch opens the file to truncate the file before closing it again. Unit test was also updated to ignore truncated PBD files in counting the number of files with data.