LocalCluster: Add ability to set hosts as removed
LocalCluster can not handle removing hosts from the cluster. This add a set of
hosts which were removed and they are skipped over during startup and not
counted in the total cluster size. It is done as a set of hosts to skip because
this keeps the hostId in the log name the same and it makes it easier for
dealing with the volt roots which use a String as the key.