ENG-15309: On rejoin send the lastSpUniqueId associated with the snapshot
During a rejoin SpInitiator.updateReplicasForJoin() calls
SpScheduler.updateReplicas() and then calls
SpScheduler.forwardPendingTaskToRejoinNode(). The first method sends the most
recently generated uniqueId to the replica which the replica might record in
the SpDurabilityListener. Then the next method will send any pending tasks to
the replica whose uniqueId will be sent to SpDurabilityListener. This can cause
the spMaxUniqueId stored in SpDurabilityListener to decrease in value.

Instead of always sending the last generated uniqueId on a rejoin send the
lastSpUniqueId associated with the snapshot which is currently being performed. All
pending transactions which are going to be sent are guaranteed to have happened
after the snap shot was initialized.