ParameterSet: Do not use commons-lang during class initialization
ParameterSet is used by the client but the client is not suppose to depend upon
commons-lang3. The client now will throw ClassNotFoundError when ParameterSet
is loaded but commons-lang3 is not on the classpath. To avoid referencing
commons-lang3 during class initialization use a new empty Object array instead
of the static instance in ArrayUtils.