ENG-12876: Remove rejoin node from zk on failure
There is a time when a the zk rejoin node exists for a rejoining host but the
host is not in zookeeper. If the host failed in this window the rejoin node
would be left in zookeeper not allowing any other hosts to rejoin.

Update MeshProber.detract(int) to also take in a zookepeer instance and call
CoreZK.removeRejoinNodeIndicatorForHost to remove the rejoin node from
zookeeper.