VoltTable: Add a strict addTable
When a volt table is added to another volt table and the schemas do not match
the error messages are very vague for debugging. This adds a new method which
adds one table into another with very strict schema checks and throws a
detailed exception with the mismatching schemas if they do not match.