VoltTable: Add a strict and efficient addTables
When a volt table is added to another volt table and the schemas do not match
the error messages are very vague for debugging. This adds a new method which
adds multipe tables into another with very strict schema checks and throws a
detailed exception with the mismatching schemas if they do not match.

Also only one extra buffer is ever allocated to hold the new rows if necessary
avoiding mutliple buffer allocations and content copies.