ENG-19511: Backport: Provide a clear error message when role is changed to xdcr
[ backport dc1535488b56f352cf14cb75a895f2462e1a6ffd ]

If the DR role of a cluster was ever updated to xdcr the error message which
was printed reported errors because export was misconfigured for DR. This error
message is confusing especially because export is reporting invalid dr changes.
Create a check in the dr validation method that checks if the role is xdcr the
cluster was correctly initialized for xdcr. Remove DR specific checks from
export configuration validation.