ENG-15435: Provide a clear error message when role is changed to xdcr
If the DR role of a cluster was ever updated to xdcr the error message which
was printed reported errors because export was misconfigured for DR. This error
message is confusing especially because export is reporting invalid dr changes.
Create a check in the dr validation method that checks if the role is xdcr the
cluster was correctly initialized for xdcr. Remove DR specific checks from
export configuration validation.