PersistentTable: Simplify isReplicatedTable vs isCatalogTableReplicated
There were two methods which implied they were testing if a table was
replicated isReplicatedTable and isCatalogTableReplicated. The first one tested
if there wasn't a partition column while the second actually returned if the
table was replicated. The results of these methods can disagree for views.

Having the two methods was confusing and resulted in asserts checking the wrong
value. All callers of isReplicatedTable didn't actually care about the
partition column state and really wanted to know if the table was marked
replicated. To simplify the methods the original implementation of
isReplicatedTable was reaplaced with the implementation of
isCatalogTableReplicated and isCatalogTableReplicated was removed.