TestSaveRestoreSysprocSuite: Fix up tests
Add a tearDown which always reverts the catalog at the end of a test so that
catalog changes from one test do not bleed into another.

Remove catch blocks which surround the test and swallows the exception, making
the test always pass.

Remove unnecessary try/catch blocks which just print the exception and fail the
test. Just let junit do its job and fail when it encounters an exception.