SnapshotRestore: Should not restore randomly partitioned views
SnapshotUtils.getTablesToSave does not return randomly partitioned views so
they shouldn't be in the snapshot. However, a snapshot can be taken with a
replicated view and the schema is modified so that the view is now over a
partitioned table so that the view is now a randomly partitioned view. When the
snapshot is restored the view which was not partitioned will be in the snapshot
and need to be skipped.