ENG-15630: Rename schemaHasNoTables to schemaHasPersistentTables
The variable does not indicate the nonexistence of all tables instead it is
just persistent tables. Invert the meaning of the boolean to be the existence
of tables instead of the absence since that is clearer to read and understand.
Also make it obvious that it is refering to persistent tables only.