PBD: Integrate extra header metadata
Add the size of the extra header metadata into the segment header so the
segment knows whether or not it has any extra metadata. Update getSchema to
getExtraHeader and add a fast exit to the method if there is no extra header
present.

Move knowledge about the format of the export schema header out of PBD and into
export.