PersistentBinaryDeque: Split offer from updateExtraHeader
Use a utility method to initialize a brand new segment that way the extra
header is always populated. Also, assign the extra header to a member variable.

Instead of providing the extra header on every offer with a boolean indicating
whether not it changed, add a method which inicates it has changed.